pyproject.toml
[project]
name = "eXwonder-backend"
version = "0.2.0"
description = "Backend for simple photos hosting with elements of socal network"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery==5.4.0",
    "dj-rest-auth==6.0.0",
    "django-cors-headers==4.5.0",
    "django-environ==0.11.2",
    "django==5.1.1",
    "djangorestframework==3.15.2",
    "drf-spectacular==0.27.2",
    "flower>=2.0.1",
    "gunicorn>=23.0.0",
    "pillow==10.4.0",
    "psycopg2-binary==2.9.9",
    "pytest-django==4.9.0",
    "pytest-factoryboy==2.7.0",
    "pytest==8.3.3",
    "pytz==2024.2",
    "redis==5.1.0",
    "ruff==0.8.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
extend-select = [
    "E",
    "F",
    "DJ",
    "I",
    "ANN",
]
ignore = ["ANN204", "ANN002", "ANN003"]
dummy-variable-rgx = "__.*"
exclude = ["manage.py", "users/migrations/*.py", "posts/migrations/*.py"]

[tool.ruff.format]
quote-style = "double"
exclude = ["manage.py", "users/migrations/*.py", "posts/migrations/*.py"]
